streamlit/requirements.txt
altair==5.2.0
asn1crypto==1.5.1
atpublic==4.0
attrs==23.1.0
blinker==1.7.0
boto3==1.33.11
botocore==1.33.11
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cryptography==41.0.7
ecdsa==0.18.0
filelock==3.13.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
s3transfer==0.8.2
six==1.16.0
smmap==5.0.1
snowflake==0.5.1
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
snowflake._legacy==0.4.0
snowflake.core==0.5.1
sortedcontainers==2.4.0
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
zstandard==0.22.0
